validation, bug fixes

+++ src/app/opdracht-detail/opdracht-detail.component.ts
@@ -21,7 +21,7 @@ import { AuthenticateService } from '../authentication/services/authenticate.ser
   styleUrls: ['./opdracht-detail.component.scss']
 })
 export class OpdrachtDetailComponent implements OnInit {
-  Opdracht: Opdracht;
+  Opdracht: Opdracht = new Opdracht(0, "", "", null, "", "", "", "", null, null, true, false, null);
   editOpdracht = false;
   newOpdracht = false;
   opdrachtForm: FormGroup;
@@ -32,9 +32,11 @@ export class OpdrachtDetailComponent implements OnInit {
   opdrachtmakers;
   tagItems = [];
   newOpdrachtId;
-
-
-  constructor(private _OpdrachtService: OpdrachtService,private _OpdrachtMakerService:OpdrachtMakerService, private _TagService: TagService, private _OpdrachtTagService: OpdrachtTagService, private route: ActivatedRoute, private fb: FormBuilder, private router: Router, private authenticateService: AuthenticateService) { }
+  submittedOpdracht = false;
+  waitOpdracht = false;
+  submittedOpdrachtNew = false;
+  waitOpdrachtNew = false;
+  constructor(private _OpdrachtService: OpdrachtService, private _OpdrachtMakerService: OpdrachtMakerService, private _TagService: TagService, private _OpdrachtTagService: OpdrachtTagService, private route: ActivatedRoute, private fb: FormBuilder, private router: Router, private authenticateService: AuthenticateService) { }
 
   ngOnInit() {
     this.route.queryParams
@@ -43,9 +45,9 @@ export class OpdrachtDetailComponent implements OnInit {
           this.newOpdracht = false;
           this._OpdrachtService.getWhereId(params.opdrachtId).subscribe(result => {
             this.Opdracht = result;
-            this._OpdrachtMakerService.getDeelnameWhereOpdrachtId(result.id).subscribe(result=>{
+            this._OpdrachtMakerService.getDeelnameWhereOpdrachtId(result.id).subscribe(result => {
               console.log(result)
-              this.opdrachtmakers=result;
+              this.opdrachtmakers = result;
             })
             this._OpdrachtTagService.getWhereBedrijfId(this.Opdracht.id).subscribe(result => {
               this.opdrachtTags = result;
@@ -53,7 +55,7 @@ export class OpdrachtDetailComponent implements OnInit {
               result.forEach(opdrachtTag => {
                 var tagObject = new TagObject(opdrachtTag.tag.naam, opdrachtTag.tag.id);
                 tagHelper.push(tagObject)
-              }); 
+              });
               this.tags = tagHelper;
             })
           });
@@ -62,7 +64,7 @@ export class OpdrachtDetailComponent implements OnInit {
           this.bedrijfId = params.bedrijfId;
           this.newOpdracht = true;
           this.editOpdracht = true;
-          this.Opdracht= new Opdracht(0,"","",params.bedrijfId,"","","","",null,null,true,false,null);
+          this.Opdracht = new Opdracht(0, "", "", params.bedrijfId, "", "", "", "", null, null, true, false, null);
         }
 
       });
@@ -75,8 +77,8 @@ export class OpdrachtDetailComponent implements OnInit {
       StraatNr: new FormControl('', Validators.required),
       Id: new FormControl('', Validators.required),
       BedrijfId: new FormControl('', Validators.required),
-      Tags: new FormControl('', Validators.required),
-      Open:new FormControl('',Validators.required)
+      Tags: new FormControl(''),
+      Open: new FormControl('', Validators.required)
     })
     this.newOpdrachtForm = this.fb.group({
       Titel: new FormControl('', Validators.required),
@@ -87,54 +89,54 @@ export class OpdrachtDetailComponent implements OnInit {
       StraatNr: new FormControl('', Validators.required),
       Id: new FormControl('', Validators.required),
       BedrijfId: new FormControl('', Validators.required),
-      Tags: new FormControl('', Validators.required),
-      Open:new FormControl('',Validators.required)
+      Tags: new FormControl(''),
+      Open: new FormControl('', Validators.required)
     })
 
     this.authenticateService.getTags().subscribe(result => {
-      result.forEach(function(item) {
+      result.forEach(function (item) {
         this.tagItems.push(new TagObject(item.naam, item.id))
       }, this)
     });
   }
-  AccepteerMaker(id:number,oldOpdrachtMaker:OpdrachtMaker){
-    const opdrachtMaker= oldOpdrachtMaker;
-    opdrachtMaker.geaccepteerd= true;
-this._OpdrachtMakerService.accepteerDeelname(id,opdrachtMaker).subscribe(result=>{ 
-  
-})
+  AccepteerMaker(id: number, oldOpdrachtMaker: OpdrachtMaker) {
+    const opdrachtMaker = oldOpdrachtMaker;
+    opdrachtMaker.geaccepteerd = true;
+    this._OpdrachtMakerService.accepteerDeelname(id, opdrachtMaker).subscribe(result => {
+
+    })
   }
-  EindigOpdracht(){
-    const opdracht:Opdracht = this.Opdracht;
+  EindigOpdracht() {
+    const opdracht: Opdracht = this.Opdracht;
     opdracht.open = false;
     opdracht.klaar = true;
     this._OpdrachtService.editOpdracht(opdracht.id, opdracht).subscribe(
-    result => {
-    });
+      result => {
+      });
   }
-  viewMaker(id: number){
-    this.router.navigate(['opdracht-maker'], { queryParams: { makerId:id } })
+  viewMaker(id: number) {
+    this.router.navigate(['opdracht-maker'], { queryParams: { makerId: id } })
   }
-  SluitOpdracht(){
-    const opdracht:Opdracht = this.Opdracht;
-    opdracht.open=false;
+  SluitOpdracht() {
+    const opdracht: Opdracht = this.Opdracht;
+    opdracht.open = false;
     this._OpdrachtService.editOpdracht(opdracht.id, opdracht).subscribe(
       result => {
       }
-    ); 
+    );
   }
-  OpenOpdracht(){
-    const opdracht:Opdracht = this.Opdracht;
-    opdracht.open=true;
+  OpenOpdracht() {
+    const opdracht: Opdracht = this.Opdracht;
+    opdracht.open = true;
     this._OpdrachtService.editOpdracht(opdracht.id, opdracht).subscribe(
       result => {
       }
-    ); 
+    );
   }
-  VerwijderMaker(id:number){
-    this._OpdrachtMakerService.deleteDeelname(id).subscribe(result=>{
-      this._OpdrachtMakerService.getDeelnameWhereOpdrachtId(this.Opdracht.id).subscribe(result=>{
-        this.opdrachtmakers=result;
+  VerwijderMaker(id: number) {
+    this._OpdrachtMakerService.deleteDeelname(id).subscribe(result => {
+      this._OpdrachtMakerService.getDeelnameWhereOpdrachtId(this.Opdracht.id).subscribe(result => {
+        this.opdrachtmakers = result;
       })
     })
   }
@@ -153,39 +155,49 @@ this._OpdrachtMakerService.accepteerDeelname(id,opdrachtMaker).subscribe(result=
       }
     );
   }
+  get o() { return this.opdrachtForm.controls; }
+  get no() { return this.newOpdrachtForm.controls; }
+
   saveChangesOpdracht() {
+    this.submittedOpdracht = true;
+    this.waitOpdracht = true;
+    if (this.opdrachtForm.invalid) {
+      this.waitOpdracht = false;
+      return;
+    }
+
     var tag = this.opdrachtForm.controls[('Tags')].value;
-  
+
 
     this._OpdrachtTagService.deleteAllWhereBedrijfId(this.opdrachtForm.controls[('Id')].value).subscribe(
       result => {
         tag.forEach(tag => {
           console.log(tag.display + " " + tag.value);
           var naamt = tag.display;
-          if(tag.value == tag.display){
-            
+          if (tag.value == tag.display) {
+
             let tag: Tag = {
-              naam : naamt,
+              naam: naamt,
               id: 0
             }
 
             this._TagService.newTag(tag).subscribe(
               result => {
 
-                let tagOpdracht: OpdrachtTag = new OpdrachtTag(0,this.opdrachtForm.controls[('Id')].value,result.id,null,null)
+                let tagOpdracht: OpdrachtTag = new OpdrachtTag(0, this.opdrachtForm.controls[('Id')].value, result.id, null, null)
                 this._OpdrachtTagService.newOpdrachtTag(tagOpdracht).subscribe();
               }
             );
           }
-          else{
-            let tagOpdracht: OpdrachtTag = new OpdrachtTag(0,this.opdrachtForm.controls[('Id')].value,tag.value,null,null)
+          else {
+            let tagOpdracht: OpdrachtTag = new OpdrachtTag(0, this.opdrachtForm.controls[('Id')].value, tag.value, null, null)
 
             this._OpdrachtTagService.newOpdrachtTag(tagOpdracht).subscribe();
           }
 
           this.editOpdracht = false;
-          
-          
+
+
         });
       }
     );
@@ -197,14 +209,20 @@ this._OpdrachtMakerService.accepteerDeelname(id,opdrachtMaker).subscribe(result=
     this._OpdrachtService.editOpdracht(this.opdrachtForm.controls['Id'].value, this.opdrachtForm.value).subscribe(
       result => {
         this.editOpdracht = false;
-
+        this.waitOpdracht = false;
       },
       err => {
         alert("opdracht bestaat al")
       }
     );
   }
-  postOpdracht(){
+  postOpdracht() {
+    this.submittedOpdrachtNew = true;
+    this.waitOpdrachtNew = true;
+    if (this.newOpdrachtForm.invalid) {
+      this.waitOpdrachtNew = false;
+      return;
+    }
     console.log(this.newOpdrachtForm.value);
     console.log(this.tags)
 
@@ -212,38 +230,39 @@ this._OpdrachtMakerService.accepteerDeelname(id,opdrachtMaker).subscribe(result=
 
     this._OpdrachtService.newOpdracht(this.newOpdrachtForm.value).subscribe(
       result => {
-        
+
         this.newOpdrachtId = result.id;
         console.log(this.newOpdrachtId);
         console.log(this.newOpdrachtForm.controls[('Tags')].value);
-      
-        var tag = this.newOpdrachtForm.controls[('Tags')].value;
-  
 
+        var tag = this.newOpdrachtForm.controls[('Tags')].value;
+
+
+        if (tag != null) {
         tag.forEach(tag => {
           console.log(tag.display + " " + tag.value);
           var naamt = tag.display;
-          if(tag.value == tag.display){
-            
+          if (tag.value == tag.display) {
+
             let tag: Tag = {
-              naam : naamt,
+              naam: naamt,
               id: 0
             }
 
             this._TagService.newTag(tag).subscribe(
               result => {
 
-                let tagOpdracht: OpdrachtTag = new OpdrachtTag(0,this.newOpdrachtId,result.id,null,null)
+                let tagOpdracht: OpdrachtTag = new OpdrachtTag(0, this.newOpdrachtId, result.id, null, null)
                 this._OpdrachtTagService.newOpdrachtTag(tagOpdracht).subscribe();
               }
             );
           }
-          else{
-            
-            let tagOpdracht: OpdrachtTag = new OpdrachtTag(0,this.newOpdrachtId,tag.value,null,null)
+          else {
+
+            let tagOpdracht: OpdrachtTag = new OpdrachtTag(0, this.newOpdrachtId, tag.value, null, null)
             console.log(tagOpdracht);
 
-            
+
             this._OpdrachtTagService.newOpdrachtTag(tagOpdracht).subscribe(
               result => {
                 console.log(result);
@@ -252,12 +271,14 @@ this._OpdrachtMakerService.accepteerDeelname(id,opdrachtMaker).subscribe(result=
           }
 
           this.editOpdracht = false;
-          
-          
+
+
         });
+      }
 
 
-          this.router.navigate(["bedrijfOpdrachten"]);
+        this.waitOpdrachtNew = false;
+        this.router.navigate(["bedrijfOpdrachten"]);
       },
       err => {
         alert("opdracht bestaat al")
